+++ 80col-mono/picoterm.h
@@ -40,13 +40,15 @@ unsigned char * slotsForRow(int y);
 void prepare_text_buffer();
 void display_terminal();
 void display_menu();
+void display_nupetscii();
 void clear_cursor();
 void print_cursor();
 void handle_new_character(unsigned char ch);
 void print_string(char str[]);
 // for menu support
 char read_key();
-char handle_menu_input();
+char handle_default_input();
+char handle_menu_input(); // specialzed for MENU_CONFIG
 bool key_ready(); // in main.c
 unsigned char read_key_from_buffer(); // in main.c
 // for debugging purposes