+++ nupetscii-font/compile_font.py
@@ -1,0 +1,220 @@
+#!/usr/bin/env python
+""" compile_font.py - generating font8-ext.C file
+
+	read the font8.c and inject the font8-ext.data into it.
+"""
+
+import datetime
+
+def get_c_def( s, name, default='' ):
+	# extract a value from a c_def string like the following
+	#    {.bitmap_index = @bitmap_index@, .adv_w = 128, .box_h = 4, .box_w = 4, .ofs_x = 2, .ofs_y = 2}
+	s = s.strip()
+	if not( ('.%s'%name) in s):
+		return default
+	_l = s.replace('{','').replace('}','').split(',')
+	for _item in _l:
+		if ('.%s'%name) in _item:
+			return _item.split('=')[1].strip()
+	return default
+
+class FontExtData:
+	""" remarks: contains the remark line. Eg: /* U+80 box (129) */
+		c_def: contains the c definition. eg: {.bitmap_index = @bitmap_index@, ... }
+		data: contains the list of bytes compiled from encoded graphic (4bpp coding) """
+
+	__slots__ = ("remark","c_def","data","__4bits" )
+
+	def __init__(self):
+		self.remark = ""
+		self.c_def = ""
+		self.data = []
+		self.__4bits = None
+
+	def add_4bits( self, value ):
+		if self.__4bits == None:
+			self.__4bits = value
+		else:
+			self.data.append( (self.__4bits << 4) + value )
+			self.__4bits = None
+
+	def extract_c_def( self, name ):
+		_r = get_c_def( self.c_def, name, None )
+		if _r==None:
+			raise Exception( 'No c_def for %s' % name )
+		return _r
+
+	@property
+	def has_4bits( self ):
+		return self.__4bits != None
+
+	def __repr__( self ):
+		return "<FontExtData \"%s\">" % self.remark.replace('/*','').replace('*/','').strip()
+
+	def print_it( self ):
+		def split_str(seq, chunk, skip_tail=False):
+			lst = []
+			if chunk <= len(seq):
+				lst.extend([seq[:chunk]])
+				lst.extend(split_str(seq[chunk:], chunk, skip_tail))
+			elif not skip_tail and seq:
+				lst.extend([seq])
+			return lst
+
+		def as_bin( i ):
+			# Transform a value into 8bits string
+			s = bin( i ).replace( '0b', '' )
+			return ('%8s' % s).replace(' ','0')
+
+		print( self ) # show representation
+		s = ''
+		for i in self.data:
+			s += as_bin( i )
+
+		# 4 Bit per Pixel
+		bpp_list = split_str( s, 4 )
+		BPP_AS_CHAR = [' ', ' ', ' ', '.', '.', '.', '.', '.', '.', '.', '.', '.', '*', '*', '*', '*']
+
+		_s = ''
+		for entry in bpp_list:
+			_s += BPP_AS_CHAR[ eval('0b%s' % entry) ]
+		box_w = int( self.extract_c_def('box_w') )
+		for substr in split_str( _s, box_w ):
+			print( substr )
+		print( '' )
+
+def load_extensions( filename ):
+	""" parse the '-ext.txt' file and return a list of FontExtData """
+	_l = []
+	with open( filename, 'r' ) as f:
+		lines = f.readlines()
+
+	_ext_data = None
+	for line in lines:
+		line = line.replace('\r','').replace('\n','')
+		if ('/*' in line) and ('*/' in line):
+			# be sure that we pushed every 4bpp to the data list.
+			if (_ext_data != None) and (_ext_data.has_4bits!=None):
+				_ext_data.add_4bits( 0x00 )
+			# Add a new entry
+			_ext_data = FontExtData()
+			_l.append( _ext_data )
+			_ext_data.remark = line.strip()
+		elif ('.bitmap_index' in line ):
+			_ext_data.c_def = line.strip()
+		else:
+			# It is a string with F....F defining entry
+			for c in line.strip():
+				if c == '.':
+					_ext_data.add_4bits( 0x0 )
+				else:
+					_ext_data.add_4bits( eval('0x%s'%c) )
+	return _l
+
+def inject_extension( font_source, font_destin, exts ):
+	_r = []
+	_r.append( '/* generated with compile_font.py on %s */\r\n' % datetime.datetime.now().strftime("%B %d, %Y  %H:%M:%S" ) )
+
+	with open( font_source ) as f:
+		_l = f.readlines()
+	# Looking for the "};" after the "/* U+7E "~" */"
+	while True:
+		line = _l.pop(0)
+		if not( "};" in line ):
+			_r.append( line )
+		else:
+			break
+
+	_r.append( '\r\n' )
+	_r.append( '/* FONT EXTENSION  added by compile_font.py */ \r\n')
+	_r.append( '\r\n' )
+
+	# Insert the extension data
+	for ext in exts:
+		_r.append( ext.remark )
+		_r.append( '\r\n' )
+		_r.append( ',' )
+		_r.append( ','.join( [hex(value) for value in ext.data]  ) )
+		_r.append( '\r\n' )
+
+	# Append the structure closure (we already read)
+	_r.append( line )
+
+	# Locate start of Glyph description section
+	while True:
+		line = _l.pop(0)
+		if 'lv_font_fmt_txt_glyph_dsc_t' in line:
+			_r.append( line )
+			break
+		_r.append( line )
+	# from now, we capture the bitmap_index value while copying the line
+	# to the destination
+	last_index = 0
+	last_box_h = 0
+	lasy_box_w = 0
+	while True:
+		line = _l.pop(0)
+		# detect the end of Glyph structure
+		if "};" in line:
+			break
+		_r.append( line )
+		# capture the last "bitmap_index" value
+		if '.bitmap_index' in line:
+			last_index = int( get_c_def( line, 'bitmap_index', -1 ) )
+			last_box_h = int( get_c_def( line, 'box_h', -1 ) )
+			last_box_w = int( get_c_def( line, 'box_w', -1 ) )
+
+	print( 'last bitmap_index = %i, box_h = %i, box_w = %i' % (last_index, last_box_h, last_box_w) )
+
+	# now we can add our extension declaration :-)
+	_r.append( '\r\n' )
+	_r.append( '/* FONT EXTENSION added by compile_font.py */ \r\n')
+	_r.append( '\r\n' )
+	for ext in exts:
+		last_index = last_index + ((last_box_h*last_box_w)/2)
+		_r.append( ',%s\r\n' % ext.c_def.replace( '@bitmap_index@', str(last_index)) )
+		last_box_h = int(ext.extract_c_def( 'box_h' ))
+		last_box_w = int(ext.extract_c_def( 'box_w' ))
+		if (last_box_h*last_box_w)%2 !=0:
+			raise Exception( "Error on %s : box_h * box_w is not a EVEN number!" % ext.remark )
+	# Close the data structure
+	_r.append( line )
+
+	# Locate the lv_font_fmt_txt_cmap_t declaration
+	# we must change the value of .range_length
+	while True:
+		line = _l.pop(0)
+		if 'range_length' in line:
+			break
+		_r.append( line )
+	# Now update the line to add the extra items we have
+	_new_items = []
+	_items = line.split(',')
+	for _item in _items:
+		if 'range_length' in _item:
+			_new_items.append( '%s+%s' % (_item, len(exts)) )
+		else:
+			_new_items.append( _item )
+	line = ','.join(_new_items) # recompose the line
+	# write the updated line
+	_r.append( line )
+
+	# We finish the source file copy
+	while len(_l)>0:
+		line = _l.pop(0)
+		if line == None:
+			break
+		_r.append( line )
+
+	# save the destination file
+	with open( font_destin, 'w' ) as f:
+		for line in _r:
+			f.write( line )
+
+
+if __name__ == '__main__':
+	exts = load_extensions( 'nupetscii.data' )
+	for item in exts:
+		item.print_it()
+	inject_extension( 'font8.c', 'nupetscii.c', exts )
+	print( 'nupetscii.c created!')